PROJECT-SOURCE-CODE/requirements.txt
opencv-python==4.6.0.66
matplotlib==3.6.2
numpy==1.23.3
easyocr==1.6.2
scikit-learn==1.1.3
pandas==1.5.1
Pillow==9.3.0
scikit-image==0.17.2
pypng==0.20220715.0
PyQRCode==1.2.1
twilio==7.16.3
mysql-connector-python==8.0.32
Flask==2.2.3